+++ src/utils/rendring/dda.c
@@ -6,7 +6,7 @@
 /*   By: melhadou <melhadou@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/01 12:01:35 by melhadou          #+#    #+#             */
-/*   Updated: 2023/12/11 17:15:53 by melhadou         ###   ########.fr       */
+/*   Updated: 2023/12/14 15:51:10 by melhadou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,12 +19,12 @@ void	dda(t_mlx mlx, t_player start, t_player end)
 	dda.dx = end.x - start.x;
 	dda.dy = end.y - start.y;
 	dda.color = end.color;
-	dda.steps = fmaxf(fabsf(dda.dx), fabsf(dda.dy));
+	dda.steps = fmaxf(fabs(dda.dx), fabs(dda.dy));
 	dda.x_inc = dda.dx / dda.steps;
 	dda.y_inc = dda.dy / dda.steps;
 	dda.x = start.x;
 	dda.y = start.y;
-	while ((int)(dda.steps--) >= 0)
+	while ((dda.steps--) >= 0)
 	{
 		my_mlx_pixel_put(mlx.img, \
 			dda.x, \
